finish field suggestion
